Refactored cpde
